angular: always extract argTypes, even without parameters.docs
