+++ code/frameworks/angular/src/client/compodoc-types.ts
