+++ code/frameworks/angular/src/client/config.ts
@@ -3,4 +3,18 @@ import './globals';
 export { render, renderToCanvas } from './render';
 export { decorateStory as applyDecorators } from './decorateStory';
 
-export const parameters = { renderer: 'angular' as const };
+import { enhanceArgTypes } from 'storybook/internal/docs-tools';
+import { ArgTypesEnhancer, Parameters } from 'storybook/internal/types';
+
+import { extractArgTypes, extractComponentDescription } from './compodoc';
+
+export const parameters: Parameters = {
+  renderer: 'angular',
+  docs: {
+    story: { inline: true },
+    extractArgTypes,
+    extractComponentDescription,
+  },
+};
+
+export const argTypesEnhancers: ArgTypesEnhancer[] = [enhanceArgTypes];
