+++ code/frameworks/angular/src/client/docs/config.ts
@@ -1,14 +1,10 @@
-import { SourceType, enhanceArgTypes } from 'storybook/internal/docs-tools';
-import { ArgTypesEnhancer, DecoratorFunction, Parameters } from 'storybook/internal/types';
+import { SourceType } from 'storybook/internal/docs-tools';
+import { DecoratorFunction, Parameters } from 'storybook/internal/types';
 
-import { extractArgTypes, extractComponentDescription } from './compodoc';
 import { sourceDecorator } from './sourceDecorator';
 
 export const parameters: Parameters = {
   docs: {
-    story: { inline: true },
-    extractArgTypes,
-    extractComponentDescription,
     source: {
       type: SourceType.DYNAMIC,
       language: 'html',
@@ -17,5 +13,3 @@ export const parameters: Parameters = {
 };
 
 export const decorators: DecoratorFunction[] = [sourceDecorator];
-
-export const argTypesEnhancers: ArgTypesEnhancer[] = [enhanceArgTypes];
