--- code/frameworks/angular/src/client/docs/index.ts
@@ -1,1 +1,0 @@
-export * from './compodoc';
