--- code/frameworks/angular/src/server/framework-preset-angular-docs.ts
@@ -1,7 +1,0 @@
-import { hasDocsOrControls } from 'storybook/internal/docs-tools';
-import { PresetProperty } from 'storybook/internal/types';
-
-export const previewAnnotations: PresetProperty<'previewAnnotations'> = (entry = [], options) => {
-  if (!hasDocsOrControls(options)) return entry;
-  return [...entry, require.resolve('../client/docs/config')];
-};
